lib,stats,dict: Make stats_dist_get_avg() return double

+++ src/lib/stats-dist.h
@@ -21,7 +21,7 @@ uint64_t stats_dist_get_min(const struct stats_dist *stats);
 /* Returns events' maximum. */
 uint64_t stats_dist_get_max(const struct stats_dist *stats);
 /* Returns events' average. */
-uint64_t stats_dist_get_avg(const struct stats_dist *stats);
+double stats_dist_get_avg(const struct stats_dist *stats);
 /* Returns events' approximate (through random subsampling) median. */
 uint64_t stats_dist_get_median(const struct stats_dist *stats);
 /* Returns events' variance */
